More changes
Moved static changes to a transformer, make PlayerInfoReplacer use the existing forge event
